+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-cli
 
+## 1.1.1
+
+### Patch Changes
+
+- 44ca7d7: refactor top level cli error handling
+- Updated dependencies [44ca7d7]
+- Updated dependencies [ca92f23]
+  - @aws-amplify/platform-core@1.0.2
+  - @aws-amplify/sandbox@1.0.6
+
 ## 1.1.0
 
 ### Minor Changes